pyproject.toml
[project]
name = "langchain-strands-aws-comparison"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "awswrangler>=3.0.0",
    "bedrock-agentcore>=1.3.1",
    "bedrock-agentcore-starter-toolkit>=0.3.2",
    "boto3>=1.42.54",
    "deepagents>=0.4.3",
    "jupyterlab>=4.5.4",
    "langchain>=1.2.10",
    "langchain-aws>=1.0.0",
    "langchain-daytona>=0.0.2",
    "langchain-mcp-adapters>=0.1.0",
    "mcp>=1.26.0",
    "pandas<3.0.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.8.0",
    "strands-agents>=1.27.0",
    "strands-agents-tools>=0.2.21",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.15.5",
]
